feat: 2170 \transactions tests...
